Allons-y Angular!
Completely gutted everything to make way for AngularJS as the
controller layer. Added Angular Foundation for whenever I get around to
UI. User log-in system is in place and working, as is displaying the
comics coming out this week and the comics specifically on your pull
list that are coming out this week. Can add comics to subscription, but
no way to take them off at the moment, as that was all just written in
a hack-y way to make sure I can do it. A lot of this code is either
unnecessary or will be changed in the future, but it needs to be
committed at some point, and this is that point.