Web Clip Layout Added, hammer.js added (but not functional), other tweaks (see description)
Tweaked the sizing of covers and text, so only two across on small
sizes and comic titles are smaller as well. Separate CSS stylesheet for
standalone mobile web clip on iOS. Added hammer.js stuff for detecting
taps, but it breaks because of the hover state stuff. Still
investigating best way to handle that.