Added work-in-progress controller layer for Comic Vine data gathering, added some comments to help those who wish to use the codebase to set up their own longbox
Marvel’s API is so limited and it’d be nice if this web application
could support more than just one publisher’s comics. As a result I
wrote a controller for getting information from Comic Vine’s API for
getting the comics coming out on a given day. This doesn’t update the
UI in any way, and if you were to try to add something from these
results to your Firebase data model everything would fall apart, but it
is what it is for the moment.